app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
sqlparse
requests
gunicorn
pytest
python-dotenv
